+++ Sources/UDeckCore/Localization/Russian.swift
@@ -67,6 +67,7 @@ struct Russian: Vocabulary {
         case .lookStrength: "Сила"
         case .lookTintColour: "Цвет заливки"
         case .lookStates: "Состояния"
+        case .lookStateAllTogether: "Настроить все вместе"
         case .stateNotchIsTheIsland: "на этом экране остров рисует чёлка"
         case .stateDropToSeparate: "Перетащи сюда, чтобы настроить отдельно"
         case .lookSharedEverywhere: "Одинаково во всех состояниях"